+++ Il2CppDumper/Il2Cpp/Il2Cpp.cs
@@ -38,7 +38,7 @@ namespace Il2CppDumper
 
         protected Il2Cpp(Stream stream) : base(stream) { }
 
-        public void SetProperties(float version, long maxMetadataUsages)
+        public void SetProperties(double version, long maxMetadataUsages)
         {
             Version = version;
             this.maxMetadataUsages = maxMetadataUsages;
@@ -48,12 +48,12 @@ namespace Il2CppDumper
         {
             if (codeRegistration != 0 && metadataRegistration != 0)
             {
-                if (Version == 24.2f)
+                if (Version == 24.2)
                 {
                     pCodeRegistration = MapVATR<Il2CppCodeRegistration>(codeRegistration);
                     if (pCodeRegistration.reversePInvokeWrapperCount > 0x30000) //TODO
                     {
-                        Version = 24.4f;
+                        Version = 24.4;
                         codeRegistration -= PointerSize * 3;
                         Console.WriteLine($"Change il2cpp version to: {Version}");
                     }
@@ -64,7 +64,7 @@ namespace Il2CppDumper
                         var genericMethodPointersCount = genericMethodTable.Max(x => x.indices.methodIndex) + 1;
                         if (pCodeRegistration.reversePInvokeWrapperCount == genericMethodPointersCount)
                         {
-                            Version = 24.3f;
+                            Version = 24.3;
                             codeRegistration -= Is32Bit ? 8u : 16u;
                             Console.WriteLine($"Change il2cpp version to: {Version}");
                         }
@@ -83,22 +83,22 @@ namespace Il2CppDumper
         public virtual void Init(ulong codeRegistration, ulong metadataRegistration)
         {
             pCodeRegistration = MapVATR<Il2CppCodeRegistration>(codeRegistration);
-            if (Version == 27f)
+            if (Version == 27)
             {
                 if (pCodeRegistration.reversePInvokeWrapperCount > 0x30000) //TODO
                 {
-                    Version = 27.1f;
+                    Version = 27.1;
                     codeRegistration -= PointerSize;
                     Console.WriteLine($"Change il2cpp version to: {Version}");
                     Console.WriteLine("CodeRegistration : {0:x}", codeRegistration);
                     pCodeRegistration = MapVATR<Il2CppCodeRegistration>(codeRegistration);
                 }
             }
-            if (Version == 24.2f)
+            if (Version == 24.2)
             {
                 if (pCodeRegistration.reversePInvokeWrapperCount > 0x30000) //TODO
                 {
-                    Version = 24.4f;
+                    Version = 24.4;
                     codeRegistration -= PointerSize * 3;
                     Console.WriteLine($"Change il2cpp version to: {Version}");
                     Console.WriteLine("CodeRegistration : {0:x}", codeRegistration);
@@ -108,7 +108,7 @@ namespace Il2CppDumper
                 {
                     if (pCodeRegistration.codeGenModules == 0) //TODO
                     {
-                        Version = 24.3f;
+                        Version = 24.3;
                         Console.WriteLine($"Change il2cpp version to: {Version}");
                         pCodeRegistration = MapVATR<Il2CppCodeRegistration>(codeRegistration);
                     }
@@ -156,7 +156,7 @@ namespace Il2CppDumper
                 types[i].Init();
                 typeDic.Add(pTypes[i], types[i]);
             }
-            if (Version >= 24.2f)
+            if (Version >= 24.2)
             {
                 var pCodeGenModules = MapVATR<ulong>(pCodeRegistration.codeGenModules, pCodeRegistration.codeGenModulesCount);
                 codeGenModules = new Dictionary<string, Il2CppCodeGenModule>(pCodeGenModules.Length, StringComparer.Ordinal);
@@ -270,7 +270,7 @@ namespace Il2CppDumper
 
         public ulong GetMethodPointer(string imageName, Il2CppMethodDefinition methodDef)
         {
-            if (Version >= 24.2f)
+            if (Version >= 24.2)
             {
                 var methodToken = methodDef.token;
                 var ptrs = codeGenModuleMethodPointers[imageName];
